// src/campaigns/order-campaign/rules/rules.enum.ts
export enum DiscountType {
  ORDER_DISCOUNT = 'ORDER_DISCOUNT',
}

export enum DiscountMode {
  PERCENT = 'PERCENT',
  AMOUNT = 'AMOUNT',
}

export enum RuleType {
  WHOLE_CART = 'WHOLE_CART',
  CART_TOTAL_CUSTOM = 'CART_TOTAL_CUSTOM',
  BULK_PURCHASE = 'BULK_PURCHASE',
  CATEGORY_DISCOUNT = 'CATEGORY_DISCOUNT',
  PRODUCT_DISCOUNT = 'PRODUCT_DISCOUNT',
  BRAND_DISCOUNT = 'BRAND_DISCOUNT',
}